avoids highlight if no snippets are requested
