+++ app/api/search.v2/buildQuery.ts
@@ -1,26 +1,7 @@
 import { SearchQuery } from 'shared/types/SearchQueryType';
 import { RequestBody } from '@elastic/elasticsearch/lib/Transport';
-import { cleanUp, searchStringMethod } from './queryHelpers';
+import { cleanUp, extractSearchParams, snippetsHighlight } from './queryHelpers';
 import { permissionsFilters } from './permissionsFilters';
-
-async function extractSearchParams(
-  query: SearchQuery
-): Promise<{
-  searchString?: string | number | undefined;
-  fullTextSearchString?: string;
-  searchMethod: string;
-}> {
-  if (query.filter && query.filter.searchString && typeof query.filter.searchString === 'string') {
-    const { searchString } = query.filter;
-    const searchTypeKey = searchString.includes(':') ? 'searchString' : 'fullTextSearchString';
-    const searchMethod = await searchStringMethod(searchString);
-    return { [searchTypeKey]: searchString, searchMethod };
-  }
-  return {
-    searchString: query.filter?.searchString,
-    searchMethod: 'query_string',
-  };
-}
 
 export const buildQuery = async (query: SearchQuery, language: string): Promise<RequestBody> => {
   const { searchString, fullTextSearchString, searchMethod } = await extractSearchParams(query);
@@ -48,19 +29,7 @@ export const buildQuery = async (query: SearchQuery, language: string): Promise<
               score_mode: 'max',
               inner_hits: {
                 _source: false,
-                highlight: {
-                  order: 'score',
-                  pre_tags: ['<b>'],
-                  post_tags: ['</b>'],
-                  encoder: 'html',
-                  number_of_fragments: 9999,
-                  type: 'fvh',
-                  fragment_size: 300,
-                  fragmenter: 'span',
-                  fields: {
-                    'fullText_*': {},
-                  },
-                },
+                ...snippetsHighlight(query),
               },
               query: {
                 [searchMethod]: {
